requirements.txt
numpy==1.16.2
matplotlib==3.0.3
ipython==7.3.0
# tensorflow==1.7.1
tensorflow-gpu==1.11.0rc2
scikit-learn==0.20.3
Cython==0.29.6
scikit-image=0.14.2